image_collection/requirements.txt
requests==2.31.0
Pillow==10.2.0
anthropiс==0.39.0
tqdm==4.66.1
python-dotenv==1.0.0